requirements.txt
absl-py==1.4.0
appdirs==1.4.4
brotlipy==0.7.0
cachetools==5.3.1
coloredlogs==15.0.1
contourpy==1.1.0
cuda-python==12.3.0
cycler==0.11.0
dnspython==2.2.1
einops==0.7.0
entmax==1.1
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h5py==3.10.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
kiwisolver==1.4.5
Mako==1.3.1
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.7.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.3.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.2
Pillow==9.0.1
platformdirs==4.1.0
prettytable==3.9.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycosat==0.6.3
Pygments==2.17.2
pyparsing==3.0.9
python-etcd==0.4.5
pytools==2023.1.1
pytorch-quantization==2.1.3
pytz==2021.3
PyYAML==6.0
pyzmq==19.0.2
requests-oauthlib==1.3.1
rich==13.7.0
safetensors==0.4.0
scipy==1.10.1
sphinx-glpi-theme==0.5
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
timm==0.9.12
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchelastic==0.2.2
torchsummary==1.5.1
torchtext==0.12.0
torchvision==0.12.0+cu113


